genius-framework/requirements.txt
Flask==2.3.3
flask-cors==4.0.0
gunicorn==21.2.0
pinecone-client==2.2.4
openai==1.3.0
python-dotenv==1.0.0
numpy==1.24.3
